requirements.txt
vosk==0.3.44
pydub==0.25.1
python-Levenshtein==0.12.2
pyspellchecker==0.7.0